requirements-dev.txt
#
# Requirements to generate the model.
#
# Note: the en_web_core_trf model is downloaded by tox.ini.
#
# --prefer-binary
# Further requirements file for testing safety.
pytest
jupyter

# Spacy GPU support.
# spacy[cuda113]
en_core_web_trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.7.3/en_core_web_trf-3.7.3-py3-none-any.whl

# Further requirements file for running models
langchain==0.1.0
langchain-community==0.0.12
qdrant-client==1.7.0
nltk==3.8.1
tiktoken==0.5.2

# Tool to upload the model to the HuggingFace Hub.
spacy-huggingface-hub==0.0.10
